+++ cligen/argcvt.nim
@@ -475,6 +475,18 @@ proc argHelp*[T](dfl: HashSet[T], a: var ArgcvtParams): seq[string]=
   argAggHelp(dflSeq, "hashset", typ, df)
   result = @[ a.argKeys, typ, df ]
 
+# Option[T]; Only simple/scalar wrapped types work well.
+when defined cgDoOptions: # To have client code in charge of std/options import,
+  import std/options      #..cligen & argcvt would need to be `include`s.
+  proc argParse*[T](dst:var Option[T], dfl:Option[T], a:var ArgcvtParams): bool=
+    var uw: T             # An unwrapped value
+    if argParse(uw, (if dfl.isSome: dfl.get else: uw), a):
+      dst = option(uw)
+      return true
+
+  proc argHelp*[T](dfl: Option[T], a: var ArgcvtParams): seq[string] =
+    @[a.argKeys, $T, (if dfl.isSome: $dfl.get else: "?")]
+
 #import tables # Tables XXX need 2D delimiting convention
 #? intsets, lists, deques, queues, etc?
 when isMainModule: